# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Treasury API
  version: 0.2.0
  description: Transfers, reserves, holds, reconciliation, and settlement operations
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/internal/recon/periods/{id}:
    get:
      summary: Get recon period
      description: Returns a single reconciliation period by ID.
      x-platform:
        kind: resource
        entity: ReconPeriod
        affectsLedger: false
      operationId: getReconPeriod
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Reconciliation period details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401": *a1
        "404": *a4
    patch:
      summary: Update recon period
      description: Updates a reconciliation period.
      x-platform:
        kind: resource
        entity: ReconPeriod
        affectsLedger: false
      operationId: updateReconPeriod
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
      responses:
        "200":
          description: Reconciliation period updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
...
  /api/internal/recon/periods/{id}/diffs:
    get:
      summary: List diffs in period
      description: Returns a paginated list of diffs for a reconciliation period.
      x-platform:
        kind: resource
        entity: ReconDiff
        affectsLedger: false
      operationId: listReconDiffs
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of diffs.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": *a1
        "403": *a2
        "404": *a4
    post:
      summary: Create diff
      description: Creates a new diff entry in a reconciliation period.
      x-platform:
        kind: resource
        entity: ReconDiff
        affectsLedger: false
      operationId: createReconDiff
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - amount
              properties:
                type:
                  type: string
                amount:
                  type: integer
                reference_id:
                  type: string
                metadata:
                  type: object
      responses:
        "201":
          description: Diff created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
  /api/internal/recon/periods/{id}/diffs/{diffId}:
    get:
      summary: Get diff
      description: Returns a single diff by ID.
      x-platform:
        kind: resource
        entity: ReconDiff
        affectsLedger: false
      operationId: getReconDiff
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - &a6
          in: path
          name: diffId
          required: true
          description: Diff ID.
          schema:
            type: string
      responses:
        "200":
          description: Diff details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401": *a1
        "404": *a4
    patch:
      summary: Update diff
      description: Updates a diff entry.
      x-platform:
        kind: resource
        entity: ReconDiff
        affectsLedger: false
      operationId: updateReconDiff
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - *a6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Diff updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
  /api/internal/recon/periods/{id}/diffs/{diffId}/explanations:
    get:
      summary: List explanations for diff
      description: Returns a paginated list of explanations for a diff.
      x-platform:
        kind: resource
        entity: ReconExplanation
        affectsLedger: false
      operationId: listReconExplanations
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - *a6
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of explanations.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": *a1
        "403": *a2
        "404": *a4
    post:
      summary: Create explanation
      description: Creates an explanation for a reconciliation diff.
      x-platform:
        kind: resource
        entity: ReconExplanation
        affectsLedger: false
      operationId: createReconExplanation
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - *a6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                category:
                  type: string
                metadata:
                  type: object
      responses:
        "201":
          description: Explanation created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
...
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...